Final commit

+++ backend/RestaurantApp/Model/ModelOrder.cs
@@ -3,7 +3,7 @@
     public class ModelOrder
     {
         public DateTime DateOfOrder { get; set; }
-        public int Bill { get; set; }
+        public double Bill { get; set; }
         public string AdditionalComment { get; set; }
         public int StateId { get; set; }
         public List<ModelOrderPosition> OrderPositions { get; set; }
